prr-flask/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
marshmallow==3.21.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
shap==0.46.0
